// extension/src-language-server/generated/ast.ts
export interface ActionUCAs extends AstNode {
    readonly $container: Model;
    action: Reference<Command>
    system: Reference<Node>
    ucas: Array<UCA>
}

export const ActionUCAs = 'ActionUCAs';

export function isActionUCAs(item: unknown): item is ActionUCAs {
    return reflection.isInstance(item, ActionUCAs);
}

export interface Command extends AstNode {
    readonly $container: VE;
    label: string
    name: string
}

export const Command = 'Command';

export function isCommand(item: unknown): item is Command {
    return reflection.isInstance(item, Command);
}

export interface ContConstraint extends AstNode {
    readonly $container: Model;
    description: string
    name: string
    refs: Array<Reference<UCA>>
}

export const ContConstraint = 'ContConstraint';

export function isContConstraint(item: unknown): item is ContConstraint {
    return reflection.isInstance(item, ContConstraint);
}

export interface Graph extends AstNode {
    readonly $container: Model;
    name: string
    nodes: Array<Node>
}

export const Graph = 'Graph';

export function isGraph(item: unknown): item is Graph {
    return reflection.isInstance(item, Graph);
}

export interface Hazard extends AstNode {
    readonly $container: Model | Hazard;
    description: string
    header: string
    name: SubID
    refs: Array<Reference<Loss>>
    subComps: Array<Hazard>
}

export const Hazard = 'Hazard';
// ...
export interface Loss extends AstNode {
    readonly $container: Model;
    description: string
    name: string
}

export const Loss = 'Loss';

export function isLoss(item: unknown): item is Loss {
    return reflection.isInstance(item, Loss);
}

export interface LossScenario extends AstNode {
    readonly $container: Model;
    description: string
    list: HazardList
    name: string
    uca: Reference<UCA>
}

export const LossScenario = 'LossScenario';

export function isLossScenario(item: unknown): item is LossScenario {
    return reflection.isInstance(item, LossScenario);
}

export interface Model extends AstNode {
    allUCAs: Array<ActionUCAs>
    controllerConstraints: Array<ContConstraint>
    controlStructure: Graph
    hazards: Array<Hazard>
    losses: Array<Loss>
    responsibilities: Array<Resps>
    safetyCons: Array<SafetyConstraint>
    scenarios: Array<LossScenario>
    systemLevelConstraints: Array<SystemConstraint>
}

export const Model = 'Model';

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model);
}

export interface Node extends AstNode {
    readonly $container: Graph;
    actions: Array<VE>
    feedbacks: Array<VE>
    label: string
    level: number
    name: string
    variables: Array<Variable>
}

export const Node = 'Node';
// ...
export interface Resps extends AstNode {
    readonly $container: Model;
    responsiblitiesForOneSystem: Array<Responsibility>
    system: Reference<Node>
}

export const Resps = 'Resps';

export function isResps(item: unknown): item is Resps {
    return reflection.isInstance(item, Resps);
}

export interface SafetyConstraint extends AstNode {
    readonly $container: Model;
    description: string
    name: string
    refs: Reference<LossScenario>
}

export const SafetyConstraint = 'SafetyConstraint';

export function isSafetyConstraint(item: unknown): item is SafetyConstraint {
    return reflection.isInstance(item, SafetyConstraint);
}

export interface SystemConstraint extends AstNode {
    readonly $container: Model | SystemConstraint;
    description: string
    header: string
    name: SubID
    refs: Array<Reference<Hazard>>
    subComps: Array<SystemConstraint>
}

export const SystemConstraint = 'SystemConstraint';
// ...
export interface Variable extends AstNode {
    readonly $container: Node;
    name: string
    values: Array<string>
}

export const Variable = 'Variable';

export function isVariable(item: unknown): item is Variable {
    return reflection.isInstance(item, Variable);
}

export interface VE extends AstNode {
    readonly $container: Node;
    comms: Array<Command>
    target: Reference<Node>
}

export const VE = 'VE';

export function isVE(item: unknown): item is VE {
    return reflection.isInstance(item, VE);
}

export type SubID = string